+++ src/simulation/env/bucket/MainGrid.h
@@ -22,19 +22,19 @@ public:
 
 	vector<int> reconstruct_path(IntVector2& startV, IntVector2& goalV, int came_from[]);
 	bool inSide(int x, int z);
-	std::vector<std::pair<int, double>*>* neighbors(const int current);
+	std::vector<std::pair<int, float>*>* neighbors(const int current);
 	double cost(const int current, const int next);
 	void debug(IntVector2& startV, IntVector2& goalV);
 	void findPath(IntVector2& startV, IntVector2& goalV);
 	void draw_grid_from(int* cameFrom, Image* image);
-	void draw_grid_cost(double* costSoFar, Image* image);
+	void draw_grid_cost(float* costSoFar, Image* image);
 	void draw_grid_path(vector<int>* path, Image* image);
 	void drawMap(Image* image);
 private:
 	IntVector2 calculateSize(int size);
-	std::vector<std::pair<int, double>*>* tempNeighbour;
+	std::vector<std::pair<int, float>*>* tempNeighbour;
 	int* came_from;
-	double* cost_so_far;
+	float* cost_so_far;
 	BucketQueue frontier;
 	int staticCounter = 0;
 	//FibHeap frontier;
